User role
Operations secured based on user role

+++ src/main/java/com/burgerjavis/BurgerJavisController.java
@@ -5,11 +5,13 @@
 
 package com.burgerjavis;
 
+import java.security.Principal;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -50,11 +52,13 @@ public class BurgerJavisController {
 	// ORDER HANDLER
 
 	/* Return order list */
+	@Secured ({"ROLE_WAITER", "ROLE_ADMIN"})
 	@RequestMapping (value = "/orders", method = RequestMethod.GET)
-	public ResponseEntity<List<Order>> getOrders() {
+	public ResponseEntity<List<Order>> getOrders(Principal principal) {
 		List<Order> orders = null;
 		try {
-			orders = (List<Order>) orderRepository.findByFinishedFalse();
+			orders = (List<Order>) orderRepository.
+					findByUsernameIgnoreCaseAndFinishedFalse(principal.getName());
 			return new ResponseEntity<List<Order>>(orders, HttpStatus.OK);
 		} catch (Exception e) {
 			return new ResponseEntity<List<Order>>(orders, HttpStatus.INTERNAL_SERVER_ERROR);
@@ -62,12 +66,17 @@ public class BurgerJavisController {
 	}
 
 	/* Return referenced order */
+	@Secured ({"ROLE_WAITER", "ROLE_ADMIN"})
 	@RequestMapping (value = "/orders/{id}", method = RequestMethod.GET)
-	public ResponseEntity<Order> getOrder(@PathVariable ("id") String id) {
+	public ResponseEntity<Order> getOrder(@PathVariable ("id") String id, Principal principal) {
 		Order order = null;
 		try {
 			order = orderRepository.findOne(id);
 			HttpStatus httpStatus = order != null ? HttpStatus.OK : HttpStatus.NOT_FOUND;
+			if(httpStatus.is2xxSuccessful()) {
+				httpStatus = order.getUsername().equalsIgnoreCase(principal.getName()) ?
+						httpStatus : HttpStatus.UNAUTHORIZED;
+			}
 			return new ResponseEntity<Order>(order, httpStatus);
 		} catch (Exception e) {
 			return new ResponseEntity<Order>(order, HttpStatus.INTERNAL_SERVER_ERROR);
@@ -75,15 +84,20 @@ public class BurgerJavisController {
 	}
 
 	/* Modify existing order */
+	@Secured ({"ROLE_WAITER", "ROLE_ADMIN"})
 	@RequestMapping (value = "/orders/{id}", method = RequestMethod.PUT)
 	public ResponseEntity<Order> modifyOrder
-			(@PathVariable ("id") String id, @RequestBody Order order) {
+			(@PathVariable ("id") String id, @RequestBody Order order, Principal principal) {
 		Order modifiedOrder = null;
 		try {
 			Order currentOrder = orderRepository.findOne(id);
 			if(currentOrder == null) {
 				// Order not found
 				return new ResponseEntity<Order>(modifiedOrder, HttpStatus.NOT_FOUND);
+			}
+			if(!currentOrder.getUsername().equalsIgnoreCase(principal.getName())) {
+				// Unauthorized user
+				return new ResponseEntity<Order>(modifiedOrder, HttpStatus.UNAUTHORIZED);
 			}
 			if(!OrderValidator.validateOrder(order)) {
 				// Order not valid
@@ -112,12 +126,17 @@ public class BurgerJavisController {
 	}
 
 	/* Delete referenced order */
+	@Secured ({"ROLE_WAITER", "ROLE_ADMIN"})
 	@RequestMapping (value = "/orders/{id}", method = RequestMethod.DELETE)
-	public ResponseEntity<Boolean> deleteOrder (@PathVariable ("id") String id) {
+	public ResponseEntity<Boolean> deleteOrder (@PathVariable ("id") String id, Principal principal) {
 		try {
 			Order currentOrder = orderRepository.findOne(id);
 			if(currentOrder == null) {
 				return new ResponseEntity<Boolean>(false, HttpStatus.NOT_FOUND);
+			}
+			if(!currentOrder.getUsername().equalsIgnoreCase(principal.getName())) {
+				// Unauthorized user
+				return new ResponseEntity<Boolean>(false, HttpStatus.UNAUTHORIZED);
 			}
 			orderRepository.delete(currentOrder);
 			return new ResponseEntity<Boolean>(true, HttpStatus.OK);
@@ -127,12 +146,17 @@ public class BurgerJavisController {
 	}
 
 	/* Create new order */
+	@Secured ({"ROLE_WAITER", "ROLE_ADMIN"})
 	@RequestMapping (value = "/orders", method = RequestMethod.POST)
-	public ResponseEntity<Order> addOrder(@RequestBody Order order) {
+	public ResponseEntity<Order> addOrder(@RequestBody Order order, Principal principal) {
 		Order newOrder = null;
 		try {
 			if(!OrderValidator.validateOrder(order)) {
 				// Order not valid
+				return new ResponseEntity<Order>(newOrder, HttpStatus.NOT_ACCEPTABLE);
+			}
+			if(!order.getUsername().equalsIgnoreCase(principal.getName())) {
+				// Incoherent data
 				return new ResponseEntity<Order>(newOrder, HttpStatus.NOT_ACCEPTABLE);
 			}
 			List<Order> orders = orderRepository.findByNameIgnoreCase(order.getName());
@@ -159,6 +183,7 @@ public class BurgerJavisController {
 	// PRODUCT HANDLER
 
 	/* Return product list */
+	@Secured ({"ROLE_WAITER", "ROLE_ADMIN"})
 	@RequestMapping (value = "/products", method = RequestMethod.GET)
 	public ResponseEntity<List<Product>> getProducts() {
 		List<Product> products = null;
@@ -171,6 +196,7 @@ public class BurgerJavisController {
 	}
 
 	/* Return referenced product */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/products/{id}", method = RequestMethod.GET)
 	public ResponseEntity<Product> getProduct(@PathVariable ("id") String id) {
 		Product product = null;
@@ -184,6 +210,7 @@ public class BurgerJavisController {
 	}
 
 	/* Modify existing product */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/products/{id}", method = RequestMethod.PUT)
 	public ResponseEntity<Product> modifyProduct
 			(@PathVariable ("id") String id, @RequestBody Product product) {
@@ -214,6 +241,7 @@ public class BurgerJavisController {
 	}
 
 	/* Delete referenced product */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/products/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Boolean> deleteProduct (@PathVariable ("id") String id) {
 		try {
@@ -229,6 +257,7 @@ public class BurgerJavisController {
 	}
 
 	/* Create new product */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/products", method = RequestMethod.POST)
 	public ResponseEntity<Product> addProduct(@RequestBody Product product) {
 		Product newProduct = null;
@@ -252,6 +281,7 @@ public class BurgerJavisController {
 	// INGREDIENT HANDLER
 
 	/* Return ingredients list */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/ingredients", method = RequestMethod.GET)
 	public ResponseEntity<List<Ingredient>> getIngredients() {
 		List<Ingredient> ingredients = null;
@@ -264,6 +294,7 @@ public class BurgerJavisController {
 	}
 
 	/* Return referenced ingredient */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/ingredients/{id}", method = RequestMethod.GET)
 	public ResponseEntity<Ingredient> getIngredient(@PathVariable ("id") String id) {
 		Ingredient ingredient = null;
@@ -277,6 +308,7 @@ public class BurgerJavisController {
 	}
 
 	/* Modify existing ingredient */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/ingredients/{id}", method = RequestMethod.PUT)
 	public ResponseEntity<Ingredient> modifyIngredient
 			(@PathVariable ("id") String id, @RequestBody Ingredient ingredient) {
@@ -307,6 +339,7 @@ public class BurgerJavisController {
 	}
 
 	/* Delete referenced ingredient */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/ingredients/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Boolean> deleteIngredient (@PathVariable ("id") String id) {
 		try {
@@ -322,6 +355,7 @@ public class BurgerJavisController {
 	}
 
 	/* Create new ingredient */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/ingredients", method = RequestMethod.POST)
 	public ResponseEntity<Ingredient> addIngredient
 			(@RequestBody Ingredient ingredient) {
@@ -345,6 +379,7 @@ public class BurgerJavisController {
 	// CATEGORY HANDLER
 
 	/* Return category list */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/categories", method = RequestMethod.GET)
 	public ResponseEntity<List<Category>> getCategories() {
 		List<Category> categories = null;
@@ -357,6 +392,7 @@ public class BurgerJavisController {
 	}
 
 	/* Return referenced category */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/categories/{id}", method = RequestMethod.GET)
 	public ResponseEntity<Category> getCategory(@PathVariable ("id") String id) {
 		Category category = null;
@@ -370,6 +406,7 @@ public class BurgerJavisController {
 	}
 
 	/* Modify existing category */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/categories/{id}", method = RequestMethod.PUT)
 	public ResponseEntity<Category> modifyCategory
 			(@PathVariable ("id") String id, @RequestBody Category category) {
@@ -406,6 +443,7 @@ public class BurgerJavisController {
 	}
 
 	/* Delete referenced category */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/categories/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Boolean> deleteCategory (@PathVariable ("id") String id) {
 		try {
@@ -421,6 +459,7 @@ public class BurgerJavisController {
 	}
 
 	/* Create new category */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/categories", method = RequestMethod.POST)
 	public ResponseEntity<Category> addCategory(@RequestBody Category category) {
 		Category newCategory = null;
@@ -443,6 +482,7 @@ public class BurgerJavisController {
 	// SUMMARY HANDLER
 
 	/* Return summary data */
+	@Secured ("ROLE_ADMIN")
 	@RequestMapping (value = "/summary", method = RequestMethod.GET)
 	public ResponseEntity<SummaryData> getSummaryData() {
 		SummaryData data = null;
@@ -466,11 +506,11 @@ public class BurgerJavisController {
 	public ResponseEntity <Credentials> getUsers(@PathVariable ("username") String username) {
 		Credentials credentials = null;
 		try {
-			List<User> users = userRepository.findByUsernameIgnoreCase(username);
-			if (users.isEmpty()) {
+			User user = userRepository.findByUsernameIgnoreCase(username);
+			if (user == null) {
 				return new ResponseEntity<Credentials>(credentials, HttpStatus.NOT_FOUND);
 			}
-			credentials = new Credentials(users.get(0));
+			credentials = new Credentials(user);
 			return new ResponseEntity<Credentials>(credentials, HttpStatus.OK);
 		} catch (Exception e) {
 			return new ResponseEntity<Credentials>(credentials, HttpStatus.INTERNAL_SERVER_ERROR);
